重新使用CubeMX生成了文件。 测试了串口，同时收发时，可以发，但是接收会出问题，进不去IDLE中断，已经写在了todo中。 以及写了未测试的云台类。 修改了部分的测试使用的代码，串口协议绑定到了串口9而非串口6等等

+++ User/Device/JC2804.hpp
@@ -32,7 +32,7 @@ public:
 
   /* 公有成员变量（外部要使用 懒得封装） */
 
-  uint8_t _device_id;
+  uint32_t _device_id;
 
 
   /* 控制指令（写入） */
